requirements/dev.txt
pytest==2.7.2
pyhamcrest==1.8.5
mock==1.3.0
pylint==1.5.4
coverage==4.0.3
coveralls==1.1
